Refactor and add functional sphinx documentation (#6)
* Improve plotting.

* Improve plotting.

* Remove some unused code

* Update installation on Gilbreth.

* Update installation on Gilbreth.

* First rough commit of sphinx documentation

* Remove (but save to another file) unused test functions. Update some docstrings.

* Updated docstrings and refactored variable names

* Fix docstrings to make sphinx less unhappy

* Rewrite sphinx documentation

* Improved sphinx docs, but still not correct

* Refactor calc_weights to gen_weights and improve sphinx api

* Fix sphinx errors and warnings

---------
