+++ dev_scripts/install_package.sh
@@ -8,6 +8,7 @@ cd ..
 pip install -r requirements.txt
 pip install -e .
 pip install -r demo/requirements.txt
+yes | conda install sphinx
 pip install -r docs/requirements.txt 
 cd dev_scripts
