--- dev_scripts/remove_package.sh
@@ -1,12 +1,0 @@
-#!/bin/bash
-# This script purges the docs and environment
-
-cd ..
-/bin/rm -r docs/build
-/bin/rm -r dist
-/bin/rm -r mbirjax.egg-info
-/bin/rm -r build
-
-pip uninstall mbirjax
-
-cd dev_scripts
