+++ docs/source/credits.rst
@@ -1,0 +1,34 @@
+Credits
+=======
+
+**MBIRJAX Development Team**
+
+The MBIR Development Team is listed below in alphabetical order:
+
+    Charles A. Bouman, Gregery T. Buzzard.
+
+**MBIRJAX Sponsors**
+
+We would like to thank the following sponsors for their financial support in the development of both this python package:
+
+    * Eli Lilly Company.
+    * Oak Ridge National Laboratory.
+
+**Citation**
+
+Please use the following Bibtex citation when referencing this software. :cite:`sreehari2016plug` (coming soon) :cite:`venkatakrishnan2013plug`
+::
+
+    @Misc {mbirjax-2024,
+    author = {Charles A Bouman and Greg T. Buzzard},
+    title =	{{V}ectorized {C}oordinate {D}escent ({VCD}) {B}ased MBIR {R}econstruction},
+    howpublished = {Software library available from \url{https://github.com/cabouman/mbirjax}},
+    year = 2024
+    }
+
+**References**
+
+.. bibliography::
+   :style: plain
+   :labelprefix: A
+   :all:
