+++ docs/source/examples.rst
@@ -1,0 +1,11 @@
+==================
+Demos and Examples
+==================
+
+**Running Included Demo Scripts**
+
+The source package contains a number of sample Python scripts in the folder
+`[mbirjax/demo] <https://github.com/cabouman/mbirjax/demo>`__.
+These scripts can be used as templates for new users.
+
+
