+++ docs/source/mbir.rst
@@ -1,0 +1,20 @@
+mbir
+----
+.. automodule:: mbirjax
+   :members: TomographyModel
+   :undoc-members:
+   :show-inheritance:
+   :member-order: bysource
+
+   .. rubric:: **Functions:**
+
+   .. autosummary::
+      ParallelBeamModel
+      TomographyModel.recon
+      TomographyModel.forward_project
+      TomographyModel.back_project
+      TomographyModel.reshape_recon
+      TomographyModel.set_params
+      TomographyModel.get_params
+      TomographyModel.gen_weights
+      TomographyModel.gen_3d_sl_phantom
