+++ docs/source/plot_utils.rst
@@ -1,0 +1,13 @@
+plot_utils
+----------
+.. automodule:: mbirjax
+   :no-index:
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :member-order: bysource
+
+   .. rubric:: **Functions:**
+
+   .. autosummary::
+      plot_utils.slice_viewer
