+++ docs/source/refs.bib
@@ -1,0 +1,48 @@
+@article{wang2016high,
+  title={High performance model based image reconstruction},
+  author={Wang, Xiao and Sabne, Amit and Kisner, Sherman and Raghunathan, Anand and Bouman, Charles and Midkiff, Samuel},
+  booktitle = {Proceedings of the 21st ACM SIGPLAN Symposium on Principles and Practice of Parallel Programming},
+  journal={ACM SIGPLAN Notices},
+  volume={51},
+  number={8},
+  numpages = {12},
+  series = {PPoPP '16},
+  year={2016},
+  doi = {10.1145/2851141.2851163},
+}
+
+@inproceedings{wang2017massively,
+  title={Massively parallel 3D image reconstruction},
+  author={Wang, Xiao and Sabne, Amit and Sakdhnagool, Putt and Kisner, Sherman J and Bouman, Charles A and Midkiff, Samuel P},
+  booktitle={Proceedings of the International Conference for High Performance Computing, Networking, Storage and Analysis},
+  articleno = {3},
+  numpages = {12},
+  series = {SC '17},
+  year={2017},
+  note = {Selected as finalist for 2017 ACM Gordon Bell Prize},
+  doi = {10.1145/3126908.3126911},
+}
+
+@inproceedings{venkatakrishnan2013plug,
+  title={Plug-and-play priors for model based reconstruction},
+  author={Venkatakrishnan, Singanallur V and Bouman, Charles A and Wohlberg, Brendt},
+  booktitle={2013 IEEE Global Conference on Signal and Information Processing},
+  pages={945--948},
+  year={2013},
+  organization={IEEE},
+  doi={10.1109/GlobalSIP.2013.6737048}
+}
+
+@article{sreehari2016plug,
+  title={Plug-and-play priors for bright field electron tomography and sparse interpolation},
+  author={Sreehari, Suhas and Venkatakrishnan, S Venkat and Wohlberg, Brendt and Buzzard, Gregery T and Drummy, Lawrence F and Simmons, Jeffrey P and Bouman, Charles A},
+  journal={IEEE Transactions on Computational Imaging},
+  volume={2},
+  number={4},
+  pages={408--423},
+  year={2016},
+  publisher={IEEE},
+  doi={10.1109/TCI.2016.2599778}
+}
+
+
