+++ mbirjax/mbir.py
@@ -11,46 +11,20 @@ import mbirjax._utils as utils
 
 class TomographyModel:
     """
-    Represents a general model for tomographic reconstruction using MBIRJAX.
-
-    Forward Model Parameters:
-        delta_det_channel (float): Detector channel spacing in ALU.
-        det_channel_offset (float): Distance in ALU from center of detector to the source-detector line along a row.
-        sigma_y (float): Sinogram additive noise standard deviation.
-
-    Reconstruction Model Parameters:
-        init_recon (jax array): 3D volume being reconstructed with shape (num_recon_slices, num_recon_rows, num_recon_cols).
-        prox_recon (jax array): 3D proximal map input with the same shape as reconstruction.
-        num_recon_rows (int): number of reconstruction rows.
-        num_recon_cols (int): number of reconstruction columns.
-        delta_pixel_recon (float): Reconstruction pixel spacing in ALU.
-        sigma_x (float): qGGMRF prior model regularization parameter.
-        sigma_p (float): Proximal map regularization parameter.
-        p (float): qGGMRF shape parameter in the range (1,2].
-        q (float): qGGMRF shape parameter in range [p,2].
-        T (float): qGGMRF threshold parameter, >0 and typically 1.0.
-        num_neighbors (int): Number of neighbors in the qGGMRF neighborhood. Possible values are {26,18,6}.
-
-    Reconstruction Parameters:
-        auto_regularize_flag (bool): Automatically adjust regularization using sharpness and snr_db
-        prior_model_type (str): Can be one of "qggmrf" or "prox_map".
-        positivity_flag (bool): True => enforces positivity constraint,
-        initialization (str): Options include "zero", "constant", or "recon" for starting from a provided reconstruction.
-        snr_db (float): Assumed signal-to-noise ratio of the input data, expressed in decibels (dB).
-        sharpness (float): Controls the reconstructed reconstruction sharpness.
-                           sharpness=0.0 is neutral; sharpness>0 increases sharpness; sharpness<0 reduces sharpness.
-        max_resolutions (int): Specifies the maximum number of grid resolutions for multi-resolution processing.
-        max_equits (float): Defines the maximum number of equivalent iterations (equits) used in reconstruction.
-        verbose (int): Controls verbosity. 0 => quiet; 1 => basic information; 2 => full debugging information.
+    Represents a general model for tomographic reconstruction using MBIRJAX. This class encapsulates the parameters and
+    methods for the forward and back projection processes required in tomographic imaging.
     """
 
     def __init__(self, angles, sinogram_shape, **kwargs):
         """
-        Create a TomographyModel object using the given angles and sinogram shape, along with other keyword arguments.
+        Initializes a TomographyModel with specific geometry and operational settings.
+
         Args:
-            angles (array_like): A 1D array of angles in radians.  The number of angles must match the number of views.
-            sinogram (jax array): 3D jax array containing sinogram with shape (num_views, num_det_rows, num_det_channels).
-            **kwargs:
+            angles (array_like): The angles at which projections should be taken, in radians.
+            sinogram_shape (tuple): The shape of the sinogram array expected (num_views, num_det_rows, num_det_channels).
+            **kwargs: Arbitrary keyword arguments for setting model parameters dynamically.
+
+        Initializes forward and back projection methods and sets up the reconstruction size and parameters.
         """
         self.params = utils.get_default_params()
         self.forward_project, self.back_project = None, None  # These are callable functions compiled in set_params
@@ -59,50 +33,68 @@ class TomographyModel:
         self.set_params(**kwargs)
 
     def compile_projectors(self):
+        """Placeholder for compiling projector methods."""
         warnings.warn('Projectors not implemented yet')
 
-    def forward_project(self, voxel_values, voxel_indices):
+    def forward_project(self, voxel_values, indices):
         """
         Forward project the given voxel values to a sinogram.
-        In the ParallelBeam case, the indices are into a flattened 2D array of shape (recon_rows, recon_cols), and the
-        projection is done using all voxels with those indices across all the slices.
+        The indices are into a flattened 2D array of shape (recon_rows, recon_cols), and the projection is done using
+        all voxels with those indices across all the slices.
+
         Args:
-            voxel_values (jax array): 2D jax array of size (len(voxel_indices), num_recon_slices) containing values of
-                    voxels to be projected.  voxel_values[i, :] contains the values for all the slices at this location
-            voxel_indices (int array): 1D integer index array of voxels to be projected
+            voxel_values (jax.numpy.DeviceArray): 2D array of voxel values to project, size (len(voxel_indices), num_recon_slices).
+            indices (numpy.ndarray): Array of indices specifying which voxels to project.
+
         Returns:
-            A jax array of shape (len(voxel_indices), num_slices)
+            jnp array: The resulting 3D sinogram after projection.
         """
-        sinogram = self.forward_project(voxel_values, voxel_indices).block_until_ready()
+        sinogram = self.forward_project(voxel_values, indices).block_until_ready()
         gc.collect()
         return sinogram
 
     def back_project(self, sinogram, indices):
         """
         Back project the given sinogram to the voxels given by the indices.
-        In the ParallelBeam case, the indices are into a flattened 2D array of shape (recon_rows, recon_cols), and the
-        projection is done onto all voxels with those indices across all the slices.
+        The indices are into a flattened 2D array of shape (recon_rows, recon_cols), and the projection is done using
+        all voxels with those indices across all the slices.
+
         Args:
-            sinogram (jax array): 3D jax array containing sinogram with shape (num_views, num_det_rows, num_det_channels).
-            indices (int array): (N_indices, num_recon_slices) integer index array of voxels to be computed in a flattened recon.
+            sinogram (jnp array): 3D jax array containing sinogram.
+            indices (jnp array): Array of indices specifying which voxels to back project.
+
         Returns:
-            A jax array of shape (len(voxel_indices), num_slices)
+            A jax array of shape (len(indices), num_slices)
         """
         recon = self.back_project(sinogram, indices).block_until_ready()
         gc.collect()
         return recon
 
     def compute_hessian_diagonal(self, weights, angles, sinogram_shape=None):
+        """
+        Computes the diagonal elements of the Hessian matrix for given weights and angles.
+
+        Args:
+            weights (jnp array): Sinogram Weights for the Hessian computation.
+            angles (jnp array): Projection angles used in the computation.
+            sinogram_shape (tuple, optional): Shape of the sinogram, defaults to None which uses internal settings.
+
+        Returns:
+            jnp array: Diagonal of the Hessian matrix with same shape as recon.
+        """
         hessian = self.compute_hessian_diagonal(weights, angles, sinogram_shape=None).block_until_ready()
         gc.collect()
         return hessian
 
     def auto_set_regularization_params(self, sinogram, weights=1):
         """
-        Sets the value of the parameters self.sigma_y, self.sigma_x, and self.sigma_p used for use in MBIR reconstruction.
+        Automatically sets the regularization parameters (self.sigma_y, self.sigma_x, and self.sigma_p) used in MBIR reconstruction based on the provided sinogram and optional weights.
+
         Args:
-            sinogram (jax array): 3D jax array containing sinogram with shape (num_views, num_det_rows, num_det_channels).
-            weights (scalar or 3D jax array): scalar value or 3D weights array with the same shape as sinogram.
+            sinogram (jnp.array): 3D jax array containing the sinogram with shape (num_views, num_det_rows, num_det_channels).
+            weights (scalar or jnp.array, optional): Scalar value or 3D weights array with the same shape as the sinogram. Defaults to 1.
+
+        The method adjusts the regularization parameters only if `auto_regularize_flag` is set to True within the model's parameters.
         """
         if self.params.auto_regularize_flag == True:
             self.auto_set_sigma_y(sinogram, weights)
@@ -111,7 +103,8 @@ class TomographyModel:
 
     def auto_set_sigma_y(self, sinogram, weights=1):
         """
-        Sets the value of the parameter self.sigma_y used for use in MBIR reconstruction.
+        Sets the value of the parameter sigma_y used for use in MBIR reconstruction.
+
         Args:
             sinogram (jax array): 3D jax array containing sinogram with shape (num_views, num_det_rows, num_det_channels).
             weights (scalar or 3D jax array): scalar value or 3D weights array with the same shape as sinogram.
@@ -226,10 +219,13 @@ class TomographyModel:
         Updates instance parameters using keyword arguments and checks for changes to critical geometry parameters.
         If a parameter key from kwargs does not exist within the instance's parameters, a NameError is raised.
         After setting parameters, it checks if key geometry-related parameters have changed and, if so, recompiles the projectors.
+
         Args:
             **kwargs: Arbitrary keyword arguments where keys are parameter names and values are the new parameter values.
+
         Raises:
             NameError: If any key provided in kwargs is not an attribute of `self.params`.
+
         Side effects:
             May recompile projectors if geometry parameters such as angles or sinogram shape are changed.
         """
@@ -291,6 +287,7 @@ class TomographyModel:
         """
         Get the values of the listed parameter names.
         Raises an exception if a parameter name is not defined in parameters.
+
         Args:
             parameter_names: String or list of strings
 
@@ -316,11 +313,14 @@ class TomographyModel:
 
     def get_voxels_at_indices(self, recon, indices):
         """
-        Get voxels at the specified indices.
+        Retrieves voxel values from a reconstruction array at specified indices.
+
         Args:
-            recon:
-            indices:
+            recon (jnp array): The 3D reconstruction array.
+            indices (jnp array): Indices for which voxel values are required.
+
         Returns:
+            numpy.ndarray or jax.numpy.DeviceArray: Array of voxel values at the specified indices.
         """
         # Get number of rows in detector
         shape = self.get_params('sinogram_shape')
@@ -337,9 +337,11 @@ class TomographyModel:
         """
         Calculate the loss function for the forward model from the error_sinogram and weights.
         The error sinogram should be error_sinogram = measured_sinogram - forward_proj(recon)
+
         Args:
             error_sinogram (jax array): 3D error sinogram with shape (num_views, num_det_rows, num_det_channels).
             weights (jax array): 3D weights array with same shape as sinogram
+
         Returns:
             [loss].
         """
@@ -355,8 +357,10 @@ class TomographyModel:
     def _get_cos_sin_angles(angles):
         """
         Take the sin and cosine of an array of num_view angles and return as a num_view x 1 jax array.
+
         Args:
             angles: array of angles
+
         Returns:
             num_view x 1 jax array containing cos and sin of the angles
         """
@@ -368,8 +372,10 @@ class TomographyModel:
     def _get_sino_indicator(sinogram):
         """
         Compute a binary function that indicates the region of sinogram support.
+
         Args:
             sinogram (jax array): 3D jax array containing sinogram with shape (num_views, num_det_rows, num_det_channels).
+
         Returns:
             (jax array): Weights used in mbircone reconstruction, with the same array shape as ``sinogram``.
         """
@@ -382,6 +388,7 @@ class TomographyModel:
         """
         Estimate the standard deviation of the reconstruction from the sinogram.  This is used to scale sigma_p and
         sigma_x in MBIR reconstruction.
+
         Args:
             sinogram (jax array): 3D jax array containing sinogram with shape (num_views, num_det_rows, num_det_channels).
         """
@@ -406,42 +413,15 @@ class TomographyModel:
         sigma_prior = (2**sharpness) * typical_img_value
         return sigma_prior
 
-    @staticmethod
-    def calc_weights(sinogram, weight_type):
-        """
-        Compute the weights used in MBIR reconstruction.
-        Args:
-            sinogram (jax array): 3D jax array containing sinogram with shape (num_views, num_det_rows, num_det_channels).
-            weight_type (string): Type of noise model used for data
-                    - weight_type = 'unweighted' => return numpy.ones(sinogram.shape).
-                    - weight_type = 'transmission' => return numpy.exp(-sinogram).
-                    - weight_type = 'transmission_root' => return numpy.exp(-sinogram/2).
-                    - weight_type = 'emission' => return 1/(numpy.absolute(sinogram) + 0.1).
-        Returns:
-            (jax array): Weights used in mbircone reconstruction, with the same array shape as ``sinogram``.
-        Raises:
-            Exception: Raised if ``weight_type`` is not one of the above options.
-        """
-        if weight_type == 'unweighted':
-            weights = jnp.ones(sinogram.shape)
-        elif weight_type == 'transmission':
-            weights = jnp.exp(-sinogram)
-        elif weight_type == 'transmission_root':
-            weights = jnp.exp(-sinogram / 2)
-        elif weight_type == 'emission':
-            weights = 1.0 / (jnp.absolute(sinogram) + 0.1)
-        else:
-            raise Exception("calc_weights: undefined weight_type {}".format(weight_type))
-
-        return weights
-
     def recon(self, sinogram, weights=1.0):
         """
         Perform MBIR reconstruction using the Multi-Granular Vector Coordinate Descent algorithm.
         This function takes care of generating its own partitions and partition sequence.
+
         Args:
             sinogram (jax array): 3D sinogram data with shape (num_views, num_det_rows, num_det_channels).
             weights (scalar or jax array): scalar or 3D positive weights with same shape as error_sinogram.
+
         Returns:
             [recon, fm_rmse]: reconstruction and array of loss for each iteration.
         """
@@ -463,11 +443,13 @@ class TomographyModel:
         """
         Perform MBIR reconstruction using the Multi-Granular Vector Coordinate Descent algorithm
         for a given set of partitions and a prescribed partition sequence.
+
         Args:
             sinogram (jax array): 3D sinogram data with shape (num_views, num_det_rows, num_det_channels).
             partitions (tuple): A collection of K partitions, with each partition being an (N_indices) integer index array of voxels to be updated in a flattened recon.
             partition_sequence (jax array): A sequence of integers that specify which partition should be used at each iteration.
             weights (scalar or jax array): scalar or 3D positive weights with same shape as error_sinogram.
+
         Returns:
             [recon, fm_rmse]: reconstruction and array of loss for each iteration.
         """
@@ -499,10 +481,10 @@ class TomographyModel:
     def vcd_partition_iteration(self, error_sinogram, recon, partition, fm_hessian, weights=1.0):
         """
         Calculate an iteration of the VCD algorithm for each subset of the partition
-
         Each iteration of the algorithm should return a better reconstructed recon. The error_sinogram should always be:
-                error_sinogram = measured_sinogram - forward_proj(recon)
+        error_sinogram = measured_sinogram - forward_proj(recon)
         where measured_sinogram is the measured sinogram and recon is the current reconstruction.
+
         Args:
             error_sinogram (jax array): 3D error sinogram with shape (num_views, num_det_rows, num_det_channels).
             partition (int array): (K, N_indices) an integer index arrays that partitions
@@ -510,6 +492,7 @@ class TomographyModel:
             recon (jax array): 3D array reconstruction with shape (num_recon_rows, num_recon_cols, num_recon_slices).
             fm_hessian (jax array): Array with same shape as recon containing diagonal of hessian for forward model loss.
             weights (scalar or jax array): scalar or 3D positive weights with same shape as error_sinogram.
+
         Returns:
             [error_sinogram, recon]: Both have the same shape as above, but are updated to reduce overall loss function.
         """
@@ -522,11 +505,10 @@ class TomographyModel:
     def vcd_subset_iteration(self, error_sinogram, recon, indices, fm_hessian, weights=1.0):
         """
         Calculate an iteration of the VCD algorithm on a single subset of the partition
-
         Each iteration of the algorithm should return a better reconstructed recon.
         The combination of (error_sinogram, recon) form a overcomplete state that make computation efficient.
         However, it is important that at each application the state should meet the constraint that:
-                error_sinogram = measured_sinogram - forward_proj(recon)
+        error_sinogram = measured_sinogram - forward_proj(recon)
         where measured_sinogram forward_proj() is whatever forward projection is being used in reconstruction.
 
         Args:
@@ -535,6 +517,7 @@ class TomographyModel:
             recon (jax array): 3D array reconstruction with shape (num_recon_rows, num_recon_cols, num_recon_slices).
             fm_hessian (jax array): Array with same shape as recon containing diagonal of hessian for forward model loss.
             weights (scalar or jax array): scalar or 3D positive weights with same shape as error_sinogram.
+
         Returns:
             [error_sinogram, recon]: Both have the same shape as above, but are updated to reduce overall loss function.
         """
@@ -596,6 +579,38 @@ class TomographyModel:
 
         return error_sinogram, recon
 
+    @staticmethod
+    def gen_weights(sinogram, weight_type):
+        """
+        Compute the weights used in MBIR reconstruction.
+
+        Args:
+            sinogram (jax array): 3D jax array containing sinogram with shape (num_views, num_det_rows, num_det_channels).
+            weight_type (string): Type of noise model used for data
+                    - weight_type = 'unweighted' => return numpy.ones(sinogram.shape).
+                    - weight_type = 'transmission' => return numpy.exp(-sinogram).
+                    - weight_type = 'transmission_root' => return numpy.exp(-sinogram/2).
+                    - weight_type = 'emission' => return 1/(numpy.absolute(sinogram) + 0.1).
+
+        Returns:
+            (jax array): Weights used in mbircone reconstruction, with the same array shape as ``sinogram``.
+
+        Raises:
+            Exception: Raised if ``weight_type`` is not one of the above options.
+        """
+        if weight_type == 'unweighted':
+            weights = jnp.ones(sinogram.shape)
+        elif weight_type == 'transmission':
+            weights = jnp.exp(-sinogram)
+        elif weight_type == 'transmission_root':
+            weights = jnp.exp(-sinogram / 2)
+        elif weight_type == 'emission':
+            weights = 1.0 / (jnp.absolute(sinogram) + 0.1)
+        else:
+            raise Exception("gen_weights: undefined weight_type {}".format(weight_type))
+
+        return weights
+
     def gen_set_of_voxel_partitions(self):
         """
         Generates a collection of voxel partitions for an array of specified partition sizes.
@@ -634,7 +649,7 @@ class TomographyModel:
                                       0:num_iterations]
         return extended_partition_sequence
 
-    def gen_3d_shepp_logan_phantom(self):
+    def gen_3d_sl_phantom(self):
         """
         Generates a 3D Shepp-Logan phantom.
         Returns:
@@ -656,9 +671,11 @@ def pm_gradient_and_hessian(delta_prime, b, sigma_x, p, q, T):
     """
     Computes the first and second derivatives of the surrogate function at a pixel for the qGGMRF prior model.
     Calculations taken from Figure 8.5 (page 119) of FCI for the qGGMRF prior model.
+
     Args:
         delta_prime (float or np.array): (batch_size, N) array of pixel differences between center and each of N neighboring pixels.
         b (float or np.array): (1,N) array of neighbor pixel weights that usually sums to 1.0.
+
     Returns:
         float or np.array: (batch_size,) array of first derivatives of the surrogate function at pixel.
         float or np.array: (batch_size,) array of second derivatives of the surrogate function at pixel.
@@ -679,6 +696,7 @@ def pm_gradient_and_hessian(delta_prime, b, sigma_x, p, q, T):
 def pm_gradient_and_hessian_at_indices(recon, indices, sigma_x, p, q, T):
     """
     Calculate the gradient and hessian at each index location in a reconstructed image using the qGGMRF prior.
+
     Args:
         recon (jax.array): 3D reconstructed image array with shape (num_recon_rows, num_recon_cols, num_recon_slices).
         indices (int array): Array of shape (N_indices, num_recon_slices) representing the indices of voxels in a flattened array to be updated.
@@ -686,6 +704,7 @@ def pm_gradient_and_hessian_at_indices(recon, indices, sigma_x, p, q, T):
         p (float): Norm parameter p in the qGGMRF prior.
         q (float): Norm parameter q in the qGGMRF prior.
         T (float): Scaling parameter in the qGGMRF prior.
+
     Returns:
         tuple: Contains two arrays (first_derivative, second_derivative) each of shape (N_indices, num_recon_slices)
                representing the gradient and Hessian values at specified indices.
@@ -709,11 +728,6 @@ def pm_gradient_and_hessian_at_indices(recon, indices, sigma_x, p, q, T):
 
     # Create a list derived from 4 neighbors, each entry in list is a 1D vector containing the recon values at that location over all slices.
     xr = [recon[row_index + offset[0], col_index + offset[1]] for offset in offsets]
-
-    # Compute differences in the slice direction (temporary placeholder).
-    # delta_prime_up_down = jnp.diff(xs, axis=-1)  # This needs further implementation based on specific requirements.
-    # delta_prime_up = jnp.pad(delta_prime_up_down, pad_width=((0, 0), (0, 1)), mode='constant', constant_values=0)
-    # delta_prime_down = jnp.pad(delta_prime_up_down, pad_width=((0, 0), (1, 0)), mode='constant', constant_values=0)
 
     # Shift slices up with zero padding
     xs_up = jnp.roll(xs, shift=-1, axis=1).at[:, -1].set(0)
@@ -770,9 +784,11 @@ def _get_rho(delta, b, sigma_x, p, q, T):
 def _get_btilde(delta_prime, b, sigma_x, p, q, T):
     """
     Compute the quadratic surrogate coefficients btilde from page 117 of FCI for the qGGMRF prior model.
+
     Args:
         delta_prime (float or np.array): (batch_size, P) array of pixel differences between center and each of P neighboring pixels.
         b (float or np.array): (1,N) array of neighbor pixel weights that usually sums to 1.0.
+
     Returns:
         float or np.array: (batch_size, P) array of surrogate coefficients btilde.
     """
@@ -795,6 +811,7 @@ def _get_btilde(delta_prime, b, sigma_x, p, q, T):
 def get_transpose(linear_map, input_shape):
     """
     Use jax to determine the transpose of a linear map.
+
     Args:
         linear_map:  [function] The linear function to be transposed
         input_shape: [ndarray] The shape of the input to the function
