+++ mbirjax/plot_utils.py
@@ -1,6 +1,5 @@
 import matplotlib.pyplot as plt
 import numpy as np
-from matplotlib import pyplot as plt
 
 
 def display_slices(phantom, sinogram, recon):
@@ -37,11 +36,11 @@ def display_slices(phantom, sinogram, recon):
 
 global slice_index, ax, fig, cbar, img, vertical_line
 
-import matplotlib.pyplot as plt
-import numpy as np
+
+from mpl_toolkits.axes_grid1 import make_axes_locatable
 
 
-def slice_viewer(data, data2=None):
+def slice_viewer(data, data2=None, title=''):
     """
     Display slices of one or two 3D image volumes with a consistent grayscale across slices.
     Allows interactive selection of slices via a draggable line on a colorbar-like axis. If two images are provided,
@@ -50,6 +49,7 @@ def slice_viewer(data, data2=None):
     Args:
         data (numpy.ndarray or jax.numpy.DeviceArray): 3D image volume with shape (height, width, depth).
         data2 (numpy.ndarray or jax.numpy.DeviceArray, optional): Second 3D image volume with the same shape as the first.
+        title (string, optional, default=''): Figure super title
 
     The function sets up a matplotlib figure with interactive controls to view different slices
     by clicking and dragging on a custom colorbar. Each slice is displayed using the same grayscale range
@@ -65,20 +65,27 @@ def slice_viewer(data, data2=None):
     def update_slice(x):
         """Update the displayed slice based on the position of the mouse click or drag on the colorbar axis."""
         global slice_index, vertical_line
-        slice_index = int(x / ax_colorbar.get_xlim()[1] * data.shape[2])
+        slice_index = int(0.5 + x / ax_colorbar.get_xlim()[1] * data.shape[2])
         vertical_line.set_xdata([slice_index, slice_index])
         redraw_fig()
 
-    def redraw_fig():
-        """Redraw the figure to update the slice and its display."""
+    def redraw_fig(show_colorbar=False):
+        """Redraw the figure to update the slice and its display.
+        The colorbar is drawn only at initialization.
+        """
         ax.clear()
         if data2 is not None:
-            ax.imshow(np.concatenate((data[:, :, slice_index], data2[:, :, slice_index]), axis=1), cmap='gray',
+            image_divider = vmax * np.ones((data.shape[0], 5))
+            im = ax.imshow(np.concatenate((data[:, :, slice_index], image_divider, data2[:, :, slice_index]), axis=1), cmap='gray',
                       vmin=vmin, vmax=vmax)
             ax.set_title(f'Slice {slice_index} Comparison')
         else:
-            ax.imshow(data[:, :, slice_index], cmap='gray', vmin=vmin, vmax=vmax)
+            im = ax.imshow(data[:, :, slice_index], cmap='gray', vmin=vmin, vmax=vmax)
             ax.set_title(f'Slice {slice_index}')
+        if show_colorbar:
+            divider = make_axes_locatable(ax)
+            cax = divider.append_axes('right', size='5%', pad=0.05)
+            fig.colorbar(im, cax=cax, orientation='vertical')
         fig.canvas.draw_idle()
 
     def on_press(event):
@@ -94,9 +101,10 @@ def slice_viewer(data, data2=None):
     # Setup the plot
     plt.ion()  # Turn on interactive mode
     fig, axes = plt.subplots(3, 1, gridspec_kw={'height_ratios': [10, 1, 0.5]})
+    fig.suptitle(title)
     ax, ax_colorbar, ax_instruction = axes
 
-    # Setup the interactive vertical line in the colorbar axis
+    # Setup the interactive vertical line in the slider
     ax_colorbar.set_xlim(0, data.shape[2])
     ax_colorbar.set_ylim(0, 1)
     vertical_line = ax_colorbar.axvline(slice_index, color='black', linewidth=4)  # Movable line
@@ -104,7 +112,7 @@ def slice_viewer(data, data2=None):
     ax_colorbar.set_yticks([])
     ax_colorbar.set_xticks([])
 
-    # Add a label below the colorbar
+    # Add a label below the slider
     ax_instruction.text(0.5, 0.5, 'Click and drag to change slice', ha='center', va='center', fontsize=10)
     ax_instruction.set_axis_off()
 
@@ -113,7 +121,7 @@ def slice_viewer(data, data2=None):
     fig.canvas.mpl_connect('motion_notify_event', on_motion)
 
     # Initial drawing
-    redraw_fig()  # Call redraw to handle initial display for single or dual images
+    redraw_fig(show_colorbar=True)  # Call redraw to handle initial display for single or dual images
     plt.show()
 
     plt.pause(0.1)  # Delay to ensure window stays open
